Cleaned up models and relationships
