+++ models/Categories.js
@@ -1,0 +1,26 @@
+const { Model, DataTypes } = require('sequelize');
+const sequelize = require('../config/connection');
+class Categories extends Model { }
+
+Categories.init(
+  {
+    id: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+      primaryKey: true,
+      autoIncrement: true,
+    },
+    name: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+  },
+  {
+    sequelize,
+    timestamps: false,
+    freezeTableName: true,
+    underscored: true,
+    modelName: 'categories'
+  }
+);
+module.exports = Categories;
