+++ models/index.js
@@ -1,37 +1,45 @@
-const { User } = require('./User');
-const { Transactions } = require('./transactions')
-const { Categories } = require('./categories')
-const { Budgets } = require('./budgets')
+const User = require('./User');
+const Transactions = require('./Transactions')
+const Categories = require('./Categories')
+const Budgets = require('./Budgets')
 
+// one to many relationship - one user, many transactions
 User.hasMany(Transactions, {
-    foreignKey: 'user_id',
-    onDelete: 'CASCADE',
-  });
-  Transactions.belongsTo(User, {
-    foreignKey: 'user_id',
-  });
+  foreignKey: 'user_id',
+  onDelete: 'CASCADE',
+});
+Transactions.belongsTo(User, {
+  foreignKey: 'user_id',
+});
 
-  Categories.hasMany(Transactions, {
-    foreignKey: 'category_id',
-    onDelete: 'CASCADE',
-  });
+// one to many - many transactions, one category
+Categories.hasMany(Transactions, {
+  foreignKey: 'category_id',
+  onDelete: 'CASCADE',
+});
+Transactions.belongsTo(Categories, {
+  foreignKey: 'category_id',
+})
 
 User.hasMany(Budgets, {
-    foreignKey: 'user_id',
-    onDelete: 'CASCADE',
-  });
-  Budgets.hasOne(Categories, {
-    foreignKey: 'category_id',
-    onDelete: 'CASCADE',
-  });
-  Categories.belongsTo(Budgets, {
-    foreignKey: 'category_id',
-    onDelete: 'CASCADE',
-  });
+  foreignKey: 'user_id',
+  onDelete: 'CASCADE',
+});
+Budgets.belongsTo(User, {
+  foreignKey: 'user_id',
+})
 
-  Categories.belongsTo(User, {
-    foreignKey: 'category_id',
-  });
+Budgets.hasOne(Categories, {
+  foreignKey: 'category_id',
+  onDelete: 'CASCADE'
+});
+Categories.belongsTo(Budgets, {
+  foreignKey: 'category_id'
+});
+
+// Categories.belongsTo(User, {
+//   foreignKey: 'category_id',
+// });
 
 
-module.exports = { User, Transactions, Categories };
+module.exports = { User, Transactions, Categories, Budgets };
